+++ torchmetrics/functional/regression/__init__.py
@@ -23,4 +23,4 @@ from torchmetrics.functional.regression.pearson import pearson_corrcoef  # noqa:
 from torchmetrics.functional.image.psnr import psnr  # noqa: F401
 from torchmetrics.functional.regression.r2 import r2_score  # noqa: F401
 from torchmetrics.functional.regression.spearman import spearman_corrcoef  # noqa: F401
-from torchmetrics.functional.regression.ssim import ssim  # noqa: F401
+from torchmetrics.functional.image.ssim import ssim  # noqa: F401