+++ torchmetrics/regression/__init__.py
@@ -20,7 +20,6 @@ from torchmetrics.regression.mean_squared_log_error import MeanSquaredLogError  
 from torchmetrics.regression.pearson import PearsonCorrcoef  # noqa: F401
 from torchmetrics.regression.r2score import R2Score  # noqa: F401
 from torchmetrics.regression.spearman import SpearmanCorrcoef  # noqa: F401
-from torchmetrics.regression.ssim import SSIM  # noqa: F401
 from torchmetrics.regression.symmetric_mean_absolute_percentage_error import (  # noqa: F401
     SymmetricMeanAbsolutePercentageError,
 )